+++ packages/lumx-react/src/components/tooltip/Tooltip.stories.tsx
@@ -29,3 +29,25 @@ export const InlineTooltip = () => (
         {' on one word.'}
     </>
 );
+
+export const MultilineTooltip = () => (
+    <>
+        <Tooltip label={'Only one sentence.'}>
+            <Button>One line</Button>
+        </Tooltip>
+        <Tooltip label={'First sentence.\nSecond sentence.\nThird sentence.\n'}>
+            <Button>Multiline</Button>
+        </Tooltip>
+    </>
+);
+
+export const EmptyTooltip = () => (
+    <>
+        <Tooltip label={''}>
+            <Button>Empty</Button>
+        </Tooltip>
+        <Tooltip label={false && 'tooltip'}>
+            <Button>Conditionnaly not displayed</Button>
+        </Tooltip>
+    </>
+);